# Conflicts:
#	src/main/java/shop/itbook/itbookshop/common/advisor/RestControllerAdvisor.java

+++ src/main/java/shop/itbook/itbookshop/coupongroup/categorycouponapply/serviec/impl/CategoryCouponApplyServiceImpl.java
@@ -35,6 +35,9 @@ public class CategoryCouponApplyServiceImpl implements CategoryCouponApplyServic
     @Override
     @Transactional
     public void cancelCategoryCouponApplyAndChangeCouponIssues(Long couponIssueNo) {
-        categoryCouponApplyRepository.deleteById(couponIssueNo);
+
+        if (categoryCouponApplyRepository.existsById(couponIssueNo)) {
+            categoryCouponApplyRepository.deleteById(couponIssueNo);
+        }
     }
 }
